Move things around, isolate midlang
