+++ Cargo.toml
@@ -1,7 +1,9 @@
 [workspace]
 
 members = [
+    "compiler",
     "mlc",
+    "mtc",
     "midlang",
     "json_frontend",
     "qbe_backend",
