+++ json_frontend/Cargo.toml
@@ -6,7 +6,8 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-midlang = { path = "../midlang" }
-
 serde = { version = "1", features = ["derive"]}
 serde_json = "1"
+
+compiler = { path = "../compiler" }
+midlang = { path = "../midlang" }
