+++ mlc/Cargo.toml
@@ -8,6 +8,6 @@ edition = "2021"
 [dependencies]
 clap = { version = "4.4.13", features = ["derive"] }
 
+compiler = { path = "../compiler" }
 json_frontend = { path = "../json_frontend" }
-midlang = { path = "../midlang" }
 qbe_backend = { "path" = "../qbe_backend" }
