+++ mtc/Cargo.toml
@@ -1,12 +1,8 @@
 [package]
-name = "qbe_backend"
+name = "mtc"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-compiler = { path = "../compiler" }
-midlang = { path = "../midlang" }
-
-ninja-writer = "0.2.0"
